// sdk/src/generated/models/SearchSortParameter.ts
/**
 * Stack Overflow for Teams API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Search sort parameter
*/
export enum SearchSortParameter {
    Relevance = 'relevance',
    Newest = 'newest',
    Active = 'active',
    Score = 'score'
}